requirements.txt
# Core dependencies
streamlit
langchain-core
langchain-groq

# Environment Variable Management
python-dotenv

# LangSmith Tracking (optional, but used in the app)
langchain